// mobile/src/types/medication.ts
export interface Medication {
    medicationId: number;
    medicationName: string;
    description: string;
    dosageInstructions: string;
    administrationMethod: string;
    sideEffects: string;
    contraindications: string;
    storageRequirements: string;
    imageUrl: string;
    status: string;
}